Source reorganization + python fixes + {component,auto}_tuples support initializers
