+++ source/timemory/components.hpp
@@ -105,282 +105,7 @@
 //
 //======================================================================================//
 
-#include "timemory/enum.h"
-#include "timemory/mpl/apply.hpp"
-
-#include <algorithm>
-#include <cctype>
-#include <string>
-#include <vector>
-
-namespace tim
-{
-//--------------------------------------------------------------------------------------//
-//
-///  description:
-///      use this function to initialize a auto_list or component_list from a list
-///      of enumerations
-//
-///  usage:
-///      using namespace tim::component;
-///      using optional_t = tim::auto_list<wall_clock, cpu_clock, cpu_util, cuda_event>;
-//
-///      auto obj = new optional_t(__FUNCTION__, __LINE__);
-///      tim::initialize(*obj, { CPU_CLOCK, CPU_UTIL });
-//
-///  typename... _ExtraArgs
-///      required because of extra "hidden" template parameters in STL containers
-//
-template <template <typename...> class _CompList, typename... _CompTypes,
-          template <typename, typename...> class _Container, typename _Intp,
-          typename... _ExtraArgs>
-void
-initialize(_CompList<_CompTypes...>&               obj,
-           const _Container<_Intp, _ExtraArgs...>& components);
-
-//--------------------------------------------------------------------------------------//
-//
-///  description:
-///      use this function to insert tools into a bundle
-//
-///  usage:
-///      using namespace tim::component;
-///      using optional_t = tim::auto_tuple<user_list_bundle>;
-//
-///      auto obj = new optional_t(__FUNCTION__, __LINE__);
-///      tim::insert(obj.get<user_list_bundle>(), { CPU_CLOCK, CPU_UTIL });
-//
-///  typename... _ExtraArgs
-///      required because of extra "hidden" template parameters in STL containers
-//
-template <size_t _Idx, typename _Type, template <size_t, typename> class _Bundle,
-          template <typename, typename...> class _Container, typename _Intp,
-          typename... _ExtraArgs>
-void
-insert(_Bundle<_Idx, _Type>& obj, const _Container<_Intp, _ExtraArgs...>& components);
-
-//--------------------------------------------------------------------------------------//
-//
-///  description:
-///      use this function to generate an array of enumerations from a list of string
-///      that can be subsequently used to initialize an auto_list or a component_list
-///
-///  usage:
-///      using namespace tim::component;
-///      using optional_t = tim::auto_list<wall_clock, cpu_clock, cpu_util, cuda_event>;
-///
-///      auto obj = new optional_t(__FUNCTION__, __LINE__);
-///      tim::initialize(*obj, tim::enumerate_components({ "cpu_clock", "cpu_util"}));
-///
-template <typename _StringT, typename... _ExtraArgs,
-          template <typename, typename...> class _Container>
-_Container<TIMEMORY_COMPONENT>
-enumerate_components(const _Container<_StringT, _ExtraArgs...>& component_names);
-
-}  // namespace tim
-
-// initialize and enumerate_components
-#include "timemory/bits/components.hpp"
-
-namespace tim
-{
-//--------------------------------------------------------------------------------------//
-//
-//                  specializations for std::initializer_list
-//
-//--------------------------------------------------------------------------------------//
-
-template <template <typename...> class _CompList, typename... _CompTypes,
-          typename _EnumT = int>
-inline void
-initialize(_CompList<_CompTypes...>& obj, std::initializer_list<_EnumT> components)
-{
-    initialize(obj, std::vector<_EnumT>(components));
-}
-
-//--------------------------------------------------------------------------------------//
-
-template <size_t _Idx, typename _Type, template <size_t, typename> class _Bundle,
-          typename _EnumT = int>
-inline void
-insert(_Bundle<_Idx, _Type>& obj, std::initializer_list<_EnumT> components)
-{
-    insert(obj, std::vector<_EnumT>(components));
-}
-
-//--------------------------------------------------------------------------------------//
-
-inline std::vector<TIMEMORY_COMPONENT>
-enumerate_components(const std::initializer_list<std::string>& component_names)
-{
-    return enumerate_components(std::vector<std::string>(component_names));
-}
-
-//--------------------------------------------------------------------------------------//
-
-template <template <typename...> class _CompList, typename... _CompTypes>
-inline void
-initialize(_CompList<_CompTypes...>& obj, std::initializer_list<std::string> components)
-{
-    initialize(obj, enumerate_components(components));
-}
-
-//--------------------------------------------------------------------------------------//
-
-template <size_t _Idx, typename _Type, template <size_t, typename> class _Bundle>
-inline void
-insert(_Bundle<_Idx, _Type>& obj, const std::initializer_list<std::string>& components)
-{
-    insert(obj, enumerate_components(components));
-}
-
-//--------------------------------------------------------------------------------------//
-
-inline std::vector<TIMEMORY_COMPONENT>
-enumerate_components(const std::string& names, const std::string& env_id = "")
-{
-    if(env_id.length() > 0)
-        return enumerate_components(tim::delimit(get_env<std::string>(env_id, names)));
-    else
-        return enumerate_components(tim::delimit(names));
-}
-
-//--------------------------------------------------------------------------------------//
-//
-//                  extra specializations for std::string
-//
-//--------------------------------------------------------------------------------------//
-//
-/// this is for initializing with a container of string
-//
-template <template <typename...> class _CompList, typename... _CompTypes,
-          typename... _ExtraArgs, template <typename, typename...> class _Container>
-inline void
-initialize(_CompList<_CompTypes...>&                     obj,
-           const _Container<std::string, _ExtraArgs...>& components)
-{
-    initialize(obj, enumerate_components(components));
-}
-
-//--------------------------------------------------------------------------------------//
-//
-/// this is for initializing with a string
-//
-template <template <typename...> class _CompList, typename... _CompTypes>
-inline void
-initialize(_CompList<_CompTypes...>& obj, const std::string& components)
-{
-    initialize(obj, enumerate_components(tim::delimit(components)));
-}
-
-//--------------------------------------------------------------------------------------//
-//
-/// this is for initializing with a container of string
-//
-template <size_t _Idx, typename _Type, template <size_t, typename> class _Bundle,
-          typename... _ExtraArgs, template <typename, typename...> class _Container>
-inline void
-insert(_Bundle<_Idx, _Type>&                         obj,
-       const _Container<std::string, _ExtraArgs...>& components)
-{
-    insert(obj, enumerate_components(components));
-}
-
-//--------------------------------------------------------------------------------------//
-//
-/// this is for initializing with a string
-//
-template <size_t _Idx, typename _Type, template <size_t, typename> class _Bundle>
-inline void
-insert(_Bundle<_Idx, _Type>& obj, const std::string& components)
-{
-    insert(obj, enumerate_components(tim::delimit(components)));
-}
-
-//--------------------------------------------------------------------------------------//
-//
-/// this is for initializing reading an environment variable, getting a string, breaking
-/// into list of components, and initializing
-//
-namespace env
-{
-//--------------------------------------------------------------------------------------//
-
-template <template <typename...> class _CompList, typename... _CompTypes,
-          typename std::enable_if<(sizeof...(_CompTypes) > 0), int>::type = 0>
-inline void
-initialize(_CompList<_CompTypes...>& obj, const std::string& env_var,
-           const std::string& default_env)
-{
-    auto env_result = tim::get_env(env_var, default_env);
-    initialize(obj, enumerate_components(tim::delimit(env_result)));
-}
-
-//--------------------------------------------------------------------------------------//
-
-template <template <typename...> class _CompList, typename... _CompTypes,
-          typename std::enable_if<(sizeof...(_CompTypes) == 0), int>::type = 0>
-inline void
-initialize(_CompList<_CompTypes...>&, const std::string&, const std::string&)
-{}
-
-//--------------------------------------------------------------------------------------//
-
-template <size_t _Idx, typename _Type, template <size_t, typename> class _Bundle>
-inline void
-insert(_Bundle<_Idx, _Type>& obj, const std::string& env_var,
-       const std::string& default_env)
-{
-    auto env_result = tim::get_env(env_var, default_env);
-    insert(obj, enumerate_components(tim::delimit(env_result)));
-}
-
-//--------------------------------------------------------------------------------------//
-
-}  // namespace env
-
-}  // namespace tim
-
-//--------------------------------------------------------------------------------------//
-/*
-#include "timemory/variadic/auto_list.hpp"
-#include "timemory/variadic/component_list.hpp"
-
-//--------------------------------------------------------------------------------------//
-
-namespace tim
-{
-template <typename... Types>
-inline typename auto_list<Types...>::init_func_t&
-auto_list<Types...>::get_initializer()
-{
-    static init_func_t _instance = [](this_type& al) {
-        static auto env_delim =
-            tim::delimit(tim::get_env<std::string>("TIMEMORY_AUTO_LIST_INIT", ""));
-        static auto enum_result = enumerate_components(env_delim);
-        for(const auto& itr : enum_result)
-            initialize(itr, al);
-        // env::initialize(al, "TIMEMORY_AUTO_LIST_INIT", "");
-    };
-    return _instance;
-}
-
-template <typename... Types>
-inline typename component_list<Types...>::init_func_t&
-component_list<Types...>::get_initializer()
-{
-    static init_func_t _instance = [](this_type& al) {
-        static auto env_delim =
-            tim::delimit(tim::get_env<std::string>("TIMEMORY_AUTO_LIST_INIT", ""));
-        static auto enum_result = enumerate_components(env_delim);
-        for(const auto& itr : enum_result)
-            initialize((TIMEMORY_COMPONENT) itr, al);
-        // env::initialize(al, "TIMEMORY_AUTO_LIST_INIT", "");
-    };
-    return _instance;
-}
-
-//--------------------------------------------------------------------------------------//
-
-}  // namespace tim
-*/
+// #include "timemory/runtime/enumerate.hpp"
+// #include "timemory/runtime/configure.hpp"
+// #include "timemory/runtime/insert.hpp"
+// #include "timemory/runtime/initialize.hpp"
