+++ source/timemory/components/general.hpp
@@ -141,7 +141,7 @@ protected:
 private:
     static std::atomic<int64_t>& get_index()
     {
-        static std::atomic<int64_t> _instance;
+        static std::atomic<int64_t> _instance(0);
         return _instance;
     }
 
@@ -209,7 +209,7 @@ protected:
 private:
     static std::atomic<int64_t>& get_index()
     {
-        static std::atomic<int64_t> _instance;
+        static std::atomic<int64_t> _instance(0);
         return _instance;
     }
 };
