+++ source/timemory/enum.h
@@ -84,7 +84,7 @@
 
 //======================================================================================//
 //
-//      Windows DLL settings
+//      Symbol override
 //
 //======================================================================================//
