+++ source/timemory/plotting.hpp
@@ -30,7 +30,6 @@
 
 #pragma once
 
-#include "timemory/bits/components.hpp"
 #include "timemory/components.hpp"
 #include "timemory/settings.hpp"
 #include "timemory/variadic/macros.hpp"
