+++ source/timemory/utility/macros.hpp
@@ -274,40 +274,32 @@
 //      extern storage singleton
 //
 #    define TIMEMORY_DECLARE_EXTERN_INIT(TYPE)                                           \
-        template <>                                                                      \
-        details::storage_singleton_t<storage<component::TYPE>>&                          \
+        extern template impl::storage_singleton_t<storage<component::TYPE>>*             \
         get_storage_singleton<storage<component::TYPE>>();                               \
-        template <>                                                                      \
-        details::storage_singleton_t<storage<component::TYPE>>&                          \
-        get_noninit_storage_singleton<storage<component::TYPE>>();                       \
         extern template class impl::storage<component::TYPE,                             \
                                             implements_storage<component::TYPE>::value>; \
-        extern template class storage<component::TYPE>;
+        extern template class storage<component::TYPE>;                                  \
+        extern template class singleton<                                                 \
+            impl::storage<component::TYPE, implements_storage<component::TYPE>::value>,  \
+            std::unique_ptr<                                                             \
+                impl::storage<component::TYPE,                                           \
+                              implements_storage<component::TYPE>::value>,               \
+                impl::storage_deleter<impl::storage<                                     \
+                    component::TYPE, implements_storage<component::TYPE>::value>>>>;
 
 #    define TIMEMORY_INSTANTIATE_EXTERN_INIT(TYPE)                                       \
-        template <>                                                                      \
-        details::storage_singleton_t<storage<component::TYPE>>&                          \
-        get_storage_singleton<storage<component::TYPE>>()                                \
-        {                                                                                \
-            using _storage_t = storage<component::TYPE>;                                 \
-            using _single_t  = details::storage_singleton_t<_storage_t>;                 \
-            static _single_t& _instance =                                                \
-                ::tim::manager::master_instance()->get_singleton<_single_t>();           \
-            return _instance;                                                            \
-        }                                                                                \
-        template <>                                                                      \
-        details::storage_singleton_t<storage<component::TYPE>>&                          \
-        get_noninit_storage_singleton<storage<component::TYPE>>()                        \
-        {                                                                                \
-            using _storage_t = storage<component::TYPE>;                                 \
-            using _single_t  = details::storage_singleton_t<_storage_t>;                 \
-            static _single_t& _instance =                                                \
-                ::tim::manager::master_instance()->get_noninit_singleton<_single_t>();   \
-            return _instance;                                                            \
-        }                                                                                \
+        template impl::storage_singleton_t<storage<component::TYPE>>*                    \
+        get_storage_singleton<storage<component::TYPE>>();                               \
         template class impl::storage<component::TYPE,                                    \
                                      implements_storage<component::TYPE>::value>;        \
-        template class storage<component::TYPE>;
+        template class storage<component::TYPE>;                                         \
+        template class singleton<                                                        \
+            impl::storage<component::TYPE, implements_storage<component::TYPE>::value>,  \
+            std::unique_ptr<                                                             \
+                impl::storage<component::TYPE,                                           \
+                              implements_storage<component::TYPE>::value>,               \
+                impl::storage_deleter<impl::storage<                                     \
+                    component::TYPE, implements_storage<component::TYPE>::value>>>>;
 
 #else
 
@@ -379,6 +371,7 @@
 #if !defined(DEBUG_PRINT_HERE)
 #    if defined(DEBUG)
 #        define DEBUG_PRINT_HERE(fmt, ...)                                               \
+            if(::tim::settings::debug())                                                 \
             printf("> [%s@'%s':%i] " fmt "...\n", __FUNCTION__, __FILE__, __LINE__,      \
                    __VA_ARGS__)
 #    else
