+++ source/timemory/variadic/auto_hybrid.hpp
@@ -66,7 +66,7 @@ public:
     using list_type_list      = typename component_type::list_type_list;
     using data_value_type     = typename component_type::data_value_type;
     using data_label_type     = typename component_type::data_label_type;
-    using init_func_t         = std::function<void(list_type&)>;
+    using init_func_t         = std::function<void(this_type&)>;
     using string_t            = std::string;
     using captured_location_t = typename component_type::captured_location_t;
 
@@ -91,7 +91,7 @@ public:
     //
     static init_func_t& get_initializer()
     {
-        static init_func_t _instance = [](list_type&) {};
+        static init_func_t _instance = [](this_type&) {};
         return _instance;
     }
 
@@ -107,15 +107,9 @@ public:
                                 bool report_at_exit = settings::destructor_report(),
                                 const _Func& _func  = this_type::get_initializer());
 
-    template <typename _Scope, typename _Func = init_func_t>
-    inline auto_hybrid(const string_t&, _Scope = _Scope{},
-                       bool         report_at_exit = settings::destructor_report(),
-                       const _Func& _func          = this_type::get_initializer());
-
-    template <typename _Func = init_func_t>
     inline explicit auto_hybrid(component_type& tmp, bool flat = settings::flat_profile(),
-                                bool report_at_exit = settings::destructor_report(),
-                                const _Func& _func  = this_type::get_initializer());
+                                bool report_at_exit = settings::destructor_report());
+
     inline ~auto_hybrid();
 
     // copy and move
@@ -248,12 +242,12 @@ auto_hybrid<_CompTuple, _CompList>::auto_hybrid(const string_t& object_tag, bool
 {
     if(m_enabled)
     {
-        _func(m_temporary_object.get_list());
+        _func(*this);
         m_temporary_object.start();
     }
 }
 
-//======================================================================================//
+//--------------------------------------------------------------------------------------//
 
 template <typename _CompTuple, typename _CompList>
 template <typename _Func>
@@ -268,36 +262,16 @@ auto_hybrid<_CompTuple, _CompList>::auto_hybrid(const captured_location_t& objec
 {
     if(m_enabled)
     {
-        _func(m_temporary_object.get_list());
+        _func(*this);
         m_temporary_object.start();
     }
 }
 
-//======================================================================================//
+//--------------------------------------------------------------------------------------//
 
 template <typename _CompTuple, typename _CompList>
-template <typename _Scope, typename _Func>
-auto_hybrid<_CompTuple, _CompList>::auto_hybrid(const string_t& object_tag, _Scope,
-                                                bool report_at_exit, const _Func& _func)
-: m_enabled(settings::enabled())
-, m_report_at_exit(report_at_exit)
-, m_temporary_object(m_enabled ? component_type(object_tag, m_enabled,
-                                                std::is_same<_Scope, scope::flat>::value)
-                               : component_type{})
-{
-    if(m_enabled)
-    {
-        _func(m_temporary_object.get_list());
-        m_temporary_object.start();
-    }
-}
-
-//======================================================================================//
-
-template <typename _CompTuple, typename _CompList>
-template <typename _Func>
 auto_hybrid<_CompTuple, _CompList>::auto_hybrid(component_type& tmp, bool flat,
-                                                bool report_at_exit, const _Func& _func)
+                                                bool report_at_exit)
 : m_enabled(true)
 , m_report_at_exit(report_at_exit)
 , m_temporary_object(tmp.clone(true, flat))
@@ -305,12 +279,11 @@ auto_hybrid<_CompTuple, _CompList>::auto_hybrid(component_type& tmp, bool flat,
 {
     if(m_enabled)
     {
-        _func(m_temporary_object.get_list());
         m_temporary_object.start();
     }
 }
 
-//======================================================================================//
+//--------------------------------------------------------------------------------------//
 
 template <typename _CompTuple, typename _CompList>
 auto_hybrid<_CompTuple, _CompList>::~auto_hybrid()
