+++ source/timemory/variadic/bits/component_tuple.hpp
@@ -57,8 +57,9 @@ inline component_tuple<Types...>::component_tuple()
 //--------------------------------------------------------------------------------------//
 //
 template <typename... Types>
+template <typename _Func>
 inline component_tuple<Types...>::component_tuple(const string_t& key, const bool& store,
-                                                  const bool& flat)
+                                                  const bool& flat, const _Func& _func)
 : m_store(store && settings::enabled())
 , m_flat(flat)
 , m_is_pushed(false)
@@ -69,14 +70,19 @@ inline component_tuple<Types...>::component_tuple(const string_t& key, const boo
 , m_key(key)
 , m_data(data_type{})
 {
-    compute_width(key);
+    if(settings::enabled())
+        _func(*this);
+    // compute_width(key);
+    set_object_prefix(m_key);
 }
 
 //--------------------------------------------------------------------------------------//
 //
 template <typename... Types>
+template <typename _Func>
 inline component_tuple<Types...>::component_tuple(const captured_location_t& loc,
-                                                  const bool& store, const bool& flat)
+                                                  const bool& store, const bool& flat,
+                                                  const _Func& _func)
 : m_store(store && settings::enabled())
 , m_flat(flat)
 , m_is_pushed(false)
@@ -87,7 +93,10 @@ inline component_tuple<Types...>::component_tuple(const captured_location_t& loc
 , m_key(loc.get_id())
 , m_data(data_type())
 {
-    compute_width(m_key);
+    if(settings::enabled())
+        _func(*this);
+    // compute_width(m_key);
+    set_object_prefix(m_key);
 }
 
 //--------------------------------------------------------------------------------------//
@@ -409,7 +418,6 @@ component_tuple<Types...>::compute_width(const string_t& _key) const
 {
     static const string_t& _prefix = get_prefix();
     output_width(_key.length() + _prefix.length() + 1);
-    set_object_prefix(_key);
 }
 
 //--------------------------------------------------------------------------------------//
@@ -428,7 +436,7 @@ inline int64_t
 component_tuple<Types...>::output_width(int64_t width)
 {
     static auto                 memorder_v = std::memory_order_relaxed;
-    static std::atomic<int64_t> _instance;
+    static std::atomic<int64_t> _instance(0);
     int64_t                     propose_width, current_width;
     auto compute = [&]() { return std::max(_instance.load(memorder_v), width); };
     while((propose_width = compute()) > (current_width = _instance.load(memorder_v)))
